Exceptions

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,15 +3,17 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseExecption;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundExecption;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -21,7 +23,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-       Product product = repository.findById(id).get();
+       Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundExecption("Recurso não encontrado"));
        return new ProductDTO(product);
     }
 
@@ -41,10 +43,14 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch(EntityNotFoundException e){
+            throw new ResourceNotFoundExecption("Recurso não encontrado");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -54,8 +60,16 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-       repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundExecption("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseExecption("Falha de integridade referencial");
+        }
     }
 }
